refactor: move indexer error to error.rs
